Add input to control publishing of preview version in build workflow
This commit introduces a new boolean input `publish_version` to the build workflow.
This input allows users to specify whether a preview version of the application should be published.
The default value for `publish_version` is true.

The "Publish preview version" step will now only run if `inputs.publish_version` is true.