Add input to control Git reference in Android release workflow
This commit introduces a new string input `ref` to the `Flutter Build Android Release` workflow.
This input allows users to specify the branch, tag, or commit to be built.
The default value for `ref` is an empty string.

The `actions/checkout@v4` step will now use the provided `inputs.ref` for the checkout.